requirements.txt
PyQt6>=6.4.0
python-docx>=0.8.11
markdown>=3.4.1
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0
PyInstaller>=5.10.0
Pillow>=9.5.0
pytest>=7.0.0
